+++ tests/test_json_functions.py
@@ -1,7 +1,7 @@
 import pytest
 import sqlalchemy as sa
 
-from postgresql_audit import jsonb_change_key_name, jsonb_merge
+from postgresql_audit import jsonb_change_key_name
 
 
 @pytest.mark.usefixtures('activity_cls', 'table_creator')
@@ -81,7 +81,7 @@ class TestJSONBChangeKeyName(object):
 
 
 @pytest.mark.usefixtures('activity_cls', 'table_creator')
-class TestJSONBMerge(object):
+class TestJSONBConcatOperator(object):
     @pytest.mark.parametrize(
         ('data', 'merge_data', 'expected'),
         (
@@ -104,44 +104,9 @@ class TestJSONBMerge(object):
     )
     def test_raw_sql(self, session, data, merge_data, expected):
         result = session.execute(
-            '''SELECT jsonb_merge(
-                '{data}'::jsonb,
-                '{merge_data}'::jsonb
-            )'''.format(
+            '''SELECT '{data}'::jsonb || '{merge_data}'::jsonb'''.format(
                 data=data,
                 merge_data=merge_data
             )
-        ).scalar()
-        assert result == expected
-
-    @pytest.mark.parametrize(
-        ('data', 'merge_data', 'expected'),
-        (
-            (
-                {"key1": 4, "key2": 3},
-                {"key1": 5, "key3": 5},
-                {"key1": 5, "key2": 3, "key3": 5}
-            ),
-            (
-                {},
-                {"key1": 5, "key3": 5},
-                {"key1": 5, "key3": 5}
-            ),
-            (
-                {"key1": 4, "key2": 3},
-                {},
-                {"key1": 4, "key2": 3}
-            ),
-        )
-    )
-    def test_sqlalchemy_function_expr(
-        self,
-        session,
-        data,
-        merge_data,
-        expected
-    ):
-        result = session.execute(
-            sa.select([jsonb_merge(data, merge_data)])
         ).scalar()
         assert result == expected